+++ lib/Datasource/ExternalFile.php
@@ -13,22 +13,26 @@ namespace OCA\Analytics\Datasource;
 
 use OCP\IL10N;
 use Psr\Log\LoggerInterface;
+use OCA\Analytics\Service\VariableService;
 
 class ExternalFile implements IDatasource
 {
     private $logger;
     private $userId;
     private $l10n;
+    private $VariableService;
 
     public function __construct(
         $userId,
         IL10N $l10n,
-        LoggerInterface $logger
+        LoggerInterface $logger,
+        VariableService $VariableService
     )
     {
         $this->userId = $userId;
         $this->l10n = $l10n;
         $this->logger = $logger;
+        $this->VariableService = $VariableService;
     }
 
     /**
@@ -97,12 +101,13 @@ class ExternalFile implements IDatasource
         $delimiter = $this->detectDelimiter($rows[0]);
 
         $header = str_getcsv($rows[0], $delimiter);
+        //ToDo: option to have datasource without header
         //$rows = array_slice($rows, 1);
 
         $data = array();
         if (count($selectedColumns) !== 0) {
+            // if only a subset of columns or fixed column values are set, they are replaced here
             $header = $this->minimizeRow($selectedColumns, $header);
-
             foreach ($rows as $row) {
                 $data[] = $this->minimizeRow($selectedColumns, str_getcsv($row, $delimiter));
             }
@@ -117,7 +122,7 @@ class ExternalFile implements IDatasource
             'header' => $header,
             'dimensions' => array_slice($header, 0, count($header) - 1),
             'data' => $data,
-            'rawdata' => $curlResult . '-' .$delimiter  ,
+            'rawdata' => $curlResult,
             'error' => ($http_code>=200 && $http_code<300) ? 0 : 'HTTP response code: '.$http_code,
         ];
     }
@@ -129,8 +134,9 @@ class ExternalFile implements IDatasource
             if (is_numeric($selectedColumn)) {
                 $rowMinimized[] = $row[$selectedColumn - 1];
             } else {
-                $rowMinimized[] = $selectedColumn;
-            }
+                // if columns contain replacement variables, they are processed here
+                $rowMinimized[] = $this->VariableService->replaceDatasourceColumns($selectedColumn);
+             }
         }
         return $rowMinimized;
     }